feat(relay): Overhaul for production-grade reliability, performance, and observability
This commit introduces a major architectural overhaul of the RMB Relay, focusing on self-healing resilience, high-load performance, and comprehensive monitoring. Core components have been re-engineered to be more robust, efficient, and transparent in a production environment.

Architectural & Reliability Overhaul:

* Federation on Redis Streams: The inter-relay federation mechanism is completely rebuilt on Redis Streams. This provides at-least-once delivery semantics and automatic recovery for failed or pending messages, replacing the previous, less reliable queue.
* Self-Healing Substrate Client: The Substrate client and event listener are now highly resilient to network failures. They feature automatic, lock-free reconnections using `arc-swap`, capped exponential backoff, and a "single-flight" pattern to prevent thundering herd scenarios during cache misses.

Performance & Efficiency:

* Local Peer Circuit-Breaker: Message routing now uses a fast, in-memory check for locally connected peers, bypassing expensive lookups and federation logic to significantly speed up local traffic.
* Binary Serialization: The Redis cache now uses `bincode` instead of `serde_json`, reducing data size and CPU overhead for serialization and deserialization.
* Batched Message ACKs: WebSocket message acknowledgments are now batched, drastically reducing Redis round-trips and improving throughput under heavy load.
* Fail-Fast Tag: A new `fail-fast` message tag provides immediate feedback to senders if a remote destination is known to be offline, avoiding unnecessary queuing and timeouts.

Networking & Protocol Modernization:

* Hyper 1.0 & HTTP/2: The entire networking stack has been upgraded to the modern Hyper 1.0 API. Federation traffic is now explicitly configured for HTTP/2 to leverage its performance benefits.
* Protocol Field Deprecation: The `federation` and `relays` fields in the `Envelope` protobuf message have been formally marked as deprecated, aligning the protocol with the new chain-based discovery mechanism.

Observability:

* Comprehensive Prometheus Metrics: New, detailed Prometheus metrics have been added across the application, tracking the Event Listener's status (reconnects, processed blocks), cache performance (hits, misses), and session evictions.

+++ src/peer/storage/redis_storage.rs
@@ -152,7 +152,7 @@ impl Storage for RedisStorage {
         }
         let mut conn = self.get_connection().await?;
         let key = BacklogKey(&backlog.uid);
-        let _: () = conn.set_ex(&key, backlog, backlog.ttl as usize).await?;
+        let _: () = conn.set_ex(&key, backlog, backlog.ttl).await?;
 
         Ok(())
     }
@@ -189,7 +189,7 @@ impl Storage for RedisStorage {
         let resp_queue = Queue::Response.as_ref();
         let queues = (req_queue, resp_queue);
 
-        let (queue, value): (String, Value) = conn.brpop(queues, 0).await?;
+        let (queue, value): (String, Value) = conn.brpop(queues, 0.0).await?;
 
         let msg: JsonMessage = if queue == req_queue {
             JsonOutgoingRequest::from_redis_value(&value)
